core_backend/requirements.txt
# Web Framework & APIs
django>=4.2.0
django-ninja>=1.1.0
django-cors-headers>=4.3.0
pydantic>=2.0.0

# Google Drive & OAuth
google-api-python-client>=2.100.0
google-auth-httplib2>=0.1.1
google-auth-oauthlib>=1.1.0

# Document Parsing
PyPDF2>=3.0.0
python-docx>=1.1.0

# AI, LLMs & Vector Database (RAG)
langchain>=0.1.0
langchain-core>=0.1.0
langchain-google-genai>=1.0.0
langchain-huggingface>=0.0.1
langchain-chroma>=0.1.0
chromadb>=0.4.24
sentence-transformers>=2.6.0